+++ packages/core/src/helpers/custom-state.ts
@@ -1,9 +1,7 @@
 import type * as postcss from 'postcss';
 import type { Diagnostics } from '../diagnostics';
-import {
-    parseSelectorWithCache,
-} from './selector';
-import type { StylableMeta } from '../stylable-processor';
+import { parseSelectorWithCache } from './selector';
+import type { StylableMeta } from '../stylable-meta';
 import type { StylableResolver } from '../stylable-resolver';
 import { valueMapping } from '../stylable-value-parsers';
 import { validateStateArgument } from '../pseudo-states';
