+++ packages/core/src/stylable-resolver.ts
@@ -1,7 +1,13 @@
 import type { FileProcessor } from './cached-process-file';
 import type { Diagnostics } from './diagnostics';
-import type { ClassSymbol, ElementSymbol, Imported } from './stylable-meta';
-import type { ImportSymbol, StylableMeta, StylableSymbol } from './stylable-processor';
+import type { StylableMeta } from './stylable-meta';
+import type {
+    ImportSymbol,
+    ClassSymbol,
+    ElementSymbol,
+    Imported,
+    StylableSymbol,
+} from './features';
 import type { StylableTransformer } from './stylable-transformer';
 import { valueMapping } from './stylable-value-parsers';
